Move enum namespace
